+++ README.md
@@ -13,8 +13,8 @@
 
 ## ✅ 구현할 기능
 - [x] 카드 추가 뷰를 구현한다.
-- [] 비밀번호는 노출되어서는 안 된다.
-  - [] 작성된 비밀번호는 *로 표시된다.
+- [x] 비밀번호는 노출되어서는 안 된다.
+  - [x] 작성된 비밀번호는 *로 표시된다.
 - [] 각 입력 항목별 유효성 검사를 진행한다.
   - [] 카드 번호와 만료일의 경우 입력할 때 자동으로 기호(-, /)가 삽입된다.
   - [] 카드 소유자 이름의 경우 입력 글자 제한이 30자이다.
