+++ app/src/main/java/woowacourse/payments/MainActivity.kt
@@ -4,42 +4,10 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.input.rememberTextFieldState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.Placeholder
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
+import woowacourse.payments.feature.AddPaymentCardScreen
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
-import woowacourse.payments.ui.theme.Gray100
-import woowacourse.payments.ui.theme.Gray200
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -55,102 +23,6 @@ class MainActivity : ComponentActivity() {
 
 @Preview(showBackground = true)
 @Composable
-fun AddPaymentCardScreen() {
-    Column(
-        modifier = Modifier.fillMaxSize()
-    ) {
-        NewCardTopBar(
-            onBackClick = {  },
-            onSaveClick = {  }
-        )
-
-        Spacer(modifier = Modifier.height(14.dp))
-
-        PaymentCard(
-            modifier = Modifier.align(Alignment.CenterHorizontally)
-        )
-
-        Spacer(modifier = Modifier.height(28.dp))
-
-        TextField("카드 번호", "0000 - 0000 - 0000 - 0000", 24)
-        TextField("만료일", "MM / YY", 190)
-        TextField("카드 소유자 이름(선택)", "카드에 표시된 이름을 입력하세요.", 24)
-        TextField("비밀번호", "0000", 190)
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun TextField(
-    label: String = "",
-    placeholder: String = "",
-    paddingEnd: Int = 0,
-) {
-    var text by remember { mutableStateOf("") }
-
-    OutlinedTextField(
-        value = text,
-        onValueChange = { text = it },
-        label = { Text(label, color = Gray200) },
-        placeholder = { Text(placeholder, color = Gray100) },
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(start = 24.dp, end = paddingEnd.dp, bottom = 18.dp)
-    )
-}
-
-@Composable
-private fun PaymentCard(
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        contentAlignment = Alignment.CenterStart,
-        modifier = modifier
-            .shadow(8.dp)
-            .size(width = 208.dp, height = 124.dp)
-            .background(
-                color = Color(0xFF333333),
-                shape = RoundedCornerShape(5.dp),
-            )
-    ) {
-        Box(
-            modifier = Modifier
-                .padding(start = 14.dp, bottom = 10.dp)
-                .size(width = 40.dp, height = 26.dp)
-                .background(
-                    color = Color(0xFFCBBA64),
-                    shape = RoundedCornerShape(4.dp),
-                )
-        )
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun NewCardTopBar(
-    onBackClick: () -> Unit,
-    onSaveClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    TopAppBar(
-        title = { Text("카드 추가") },
-        navigationIcon = {
-            IconButton(onClick = { onBackClick() }) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = "뒤로 가기",
-                )
-            }
-        },
-        actions = {
-            IconButton(onClick = { onSaveClick() }) {
-                Icon(
-                    imageVector = Icons.Filled.Check,
-                    contentDescription = "완료",
-                )
-            }
-        },
-        modifier = modifier
-    )
+private fun PreviewAddPaymentCardScreen() {
+    AddPaymentCardScreen()
 }
